myapp/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.10.1
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
whitenoise==3.3.1
